% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{annotation_from_merged_peaks}
\alias{annotation_from_merged_peaks}
\title{Find nearest peaks of each gene and return refined annotation}
\usage{
annotation_from_merged_peaks(scExp, odir, merged_peaks, geneTSS_annotation)
}
\arguments{
\item{scExp}{A SingleCellExperiment object}

\item{odir}{An output directory where to write the mergedpeaks BED file}

\item{merged_peaks}{A list of GRanges object containing the merged peaks}

\item{geneTSS_annotation}{A GRanges object with reference genes}
}
\value{
A data.frame with refined annotation
}
\description{
Find nearest peaks of each gene and return refined annotation
}
