% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{annotToCol2}
\alias{annotToCol2}
\title{annotToCol2}
\usage{
annotToCol2(
  annotS = NULL,
  annotT = NULL,
  missing = c("", NA),
  anotype = NULL,
  maxnumcateg = 2,
  categCol = NULL,
  quantitCol = NULL,
  plotLegend = TRUE,
  plotLegendFile = NULL
)
}
\arguments{
\item{annotS}{}

\item{annotT}{}

\item{missing}{}

\item{anotype}{}

\item{maxnumcateg}{}

\item{categCol}{}

\item{quantitCol}{}

\item{plotLegend}{}

\item{plotLegendFile}{}
}
\value{
A matrix of continuous or discrete colors
}
\description{
annotToCol2
}
\examples{
data("scExp")
annotToCol2(colData(scExp))

}
