% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{anocol_binary}
\alias{anocol_binary}
\title{Helper binary column for anocol function}
\usage{
anocol_binary(anocol, anotype, plotLegend, annotS)
}
\arguments{
\item{anocol}{The color feature matrix}

\item{anotype}{The feature types}

\item{plotLegend}{Plot legend ?}

\item{annotS}{A color matrix}
}
\value{
A color matrix similar to anocol with binrary columns colored
}
\description{
Helper binary column for anocol function
}
