% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{beds_to_matrix_indexes}
\alias{beds_to_matrix_indexes}
\title{Count bed files on interval to create count indexes}
\usage{
beds_to_matrix_indexes(files_dir, which)
}
\value{
feature_indexes
}
\description{
Count bed files on interval to create count indexes
}
