% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{call_macs2_merge_peaks}
\alias{call_macs2_merge_peaks}
\title{Calling MACS2 peak caller and merging resulting peaks}
\usage{
call_macs2_merge_peaks(
  affectation,
  odir,
  p.value,
  format = c("scBED", "scBAM")[1],
  ref,
  peak_distance_to_merge
)
}
\arguments{
\item{affectation}{Annotation data.frame with cell cluster and cell id
information}

\item{odir}{Output directory to write MACS2 output}

\item{p.value}{P value to detect peaks, passed to MACS2}

\item{format}{File format, either "scBAM" or "scBED"}

\item{ref}{Reference genome to get chromosome information from.}

\item{peak_distance_to_merge}{Distance to merge peaks}
}
\value{
A list of merged GRanges peaks
}
\description{
Calling MACS2 peak caller and merging resulting peaks
}
