% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce11.GeneTSS}
\alias{ce11.GeneTSS}
\title{Data.frame of gene TSS - ce11}
\format{
ce11.GeneTSS - a data frame with 20,051 genes and 5 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the gene (TSS) - integer}
\item{end}{End of the gene - integer}
\item{Gene}{Gene symbol - character}
\item{strand}{Srand - character}
}
}
\usage{
data("ce11.GeneTSS")
}
\value{
ce11.GeneTSS - a data frame with 20,051 genes and 5 variables
}
\description{
This dataframe was extracted from Gencode v25 and report the Transcription
Start Site of each gene in the C Elegans genome build ce11
}
\keyword{datasets}
