% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{changeRange}
\alias{changeRange}
\title{changeRange}
\usage{
changeRange(v, newmin = 1, newmax = 10)
}
\arguments{
\item{v}{}

\item{newmin}{}

\item{newmax}{}
}
\value{
A matrix with values scaled between newmin and newmax
}
\description{
changeRange
}
\examples{
mat = matrix(runif(100),runif(100),ncol=1,nrow=100)
mat. = changeRange(mat,0,1)
summary(mat.)
}
