% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{check_correct_datamatrix}
\alias{check_correct_datamatrix}
\title{Check if matrix rownames are well formated and correct if needed}
\usage{
check_correct_datamatrix(datamatrix_single, sample_name = "")
}
\arguments{
\item{datamatrix_single}{A sparse matrix}

\item{sample_name}{Matrix sample name for warnings}
}
\value{
A sparseMatrix in the right rownames format
}
\description{
Throws warnings / error if matrix is in the wrong format
}
\examples{
check_correct_datamatrix(create_scDataset_raw()$mat)
}
