% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{choose_cluster_scExp}
\alias{choose_cluster_scExp}
\title{Choose a number of clusters}
\usage{
choose_cluster_scExp(scExp, nclust = 3, consensus = T, hc_linkage = "ward.D")
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing consclust in metadata.}

\item{nclust}{Number of cluster to pick [3]}

\item{hc_linkage}{A linkage method for hierarchical clustering. See \link[stats]{cor}. ['ward.D']}
}
\value{
Returns a SingleCellExperiment object with each cell assigned to a correlation cluster in colData.
}
\description{
This functions takes as input a SingleCellExperiment object with consclust and a number of cluster to select.
It outputs a SingleCellExperiment object with each cell assigned to a correlation cluster in colData. Also 
calculates a hierarchical clustering of the consensus associations calculated by ConsensusClusterPlus.
}
