% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{choose_perplexity}
\alias{choose_perplexity}
\title{Choose perplexity depending on number of cells for Tsne}
\usage{
choose_perplexity(dataset)
}
\arguments{
\item{dataset}{A matrix of features x cells (rows x columns)}
}
\value{
A number between 5 and 30 to use in Rtsne function
}
\description{
Choose perplexity depending on number of cells for Tsne
}
\examples{
choose_perplexity(matrix(runif(100),nrow = 500,ncol=500))
choose_perplexity(matrix(runif(100),nrow = 250,ncol=500))
choose_perplexity(matrix(runif(100),nrow = 50,ncol=500))
}
