% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{colors_scExp}
\alias{colors_scExp}
\title{Adding colors to cells & features}
\usage{
colors_scExp(
  scExp,
  annotCol = "sample_id",
  color_by = "sample_id",
  color_df = NULL
)
}
\arguments{
\item{scExp}{}

\item{annotCol}{}

\item{color_by}{}

\item{color_df}{}
}
\value{

}
\description{
Adding colors to cells & features
}
