% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{combine_datamatrix}
\alias{combine_datamatrix}
\title{Combine two matrices and emit warning if no regions are in common}
\usage{
combine_datamatrix(datamatrix, datamatrix_single, file_names, i)
}
\arguments{
\item{datamatrix}{A sparse matrix or NULL if empty}

\item{datamatrix_single}{Another sparse matrix}

\item{file_names}{File name corresponding to the matrix for warnings}

\item{i}{file number}
}
\value{
A combined sparse matrix
}
\description{
Combine two matrices and emit warning if no regions are in common
}
