% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{combine_enrichmentTests}
\alias{combine_enrichmentTests}
\title{Run enrichment tests and combine into list}
\usage{
combine_enrichmentTests(
  diff,
  qval.th,
  cdiff.th,
  annotFeat_long,
  peak_distance,
  refined_annotation
)
}
\arguments{
\item{diff}{Differential list}

\item{qval.th}{Differential analysis adjusted p.value threshold}

\item{cdiff.th}{Differential analysis log-fold change threshold}

\item{annotFeat_long}{Long annotation}

\item{peak_distance}{Maximum gene to peak distance}

\item{refined_annotation}{Refined annotation data.frame if peak calling is
done}
}
\value{

}
\description{
Run enrichment tests and combine into list
}
