% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{comparable_variables}
\alias{comparable_variables}
\title{Find comparable variable scExp}
\usage{
comparable_variables(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment}
}
\value{
A character vector with the comparable variable names
}
\description{
Find comparable variable scExp
}
