% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{CompareedgeRGLM}
\alias{CompareedgeRGLM}
\title{Creates a summary table with the number of genes under- or overexpressed in each group and outputs several graphical representations}
\usage{
CompareedgeRGLM(
  dataMat = NULL,
  annot = NULL,
  ref = NULL,
  groups = NULL,
  featureTab = NULL,
  norm_method = "RLE"
)
}
\arguments{
\item{dataMat}{reads matrix}

\item{annot}{selected annotation of interest}

\item{ref}{List containing one or more vectors of reference samples. Name of the vectors will be used in the results table. The length of this list should be 1 or the same length as the groups list}

\item{groups}{List containing the IDs of groups to be compared with the reference samples. Names of the vectors will be used in the results table}

\item{featureTab}{Feature annotations to be added to the results table}

\item{norm_method}{}
}
\value{
Results table
}
\description{
Creates a summary table with the number of genes under- or overexpressed in each group and outputs several graphical representations
}
\author{
Eric Letouze & Celine Vallot
}
