% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{concatenate_scBed_into_clusters}
\alias{concatenate_scBed_into_clusters}
\title{Concatenate single-cell BED into clusters}
\usage{
concatenate_scBed_into_clusters(affectation, files, odir, ref)
}
\arguments{
\item{affectation}{Annotation data.frame containing cluster information}

\item{files}{Input files to concatenate (must match affectation$barcode)}

\item{odir}{Output directory}
}
\value{
Merge single-cell BED files into cluster BED files. Ungzip file if
BED is gzipped.
}
\description{
Concatenate single-cell BED into clusters
}
