% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{create_sample_name_mat}
\alias{create_sample_name_mat}
\title{Create a sample name matrix}
\usage{
create_sample_name_mat(nb_samples, samples_names)
}
\arguments{
\item{nb_samples}{Number of samples}

\item{samples_names}{Character vector of sample names}
}
\value{
A matrix
}
\description{
Create a sample name matrix
}
