% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{create_scExp}
\alias{create_scExp}
\title{Title}
\usage{
create_scExp(
  datamatrix,
  annot,
  removeZeroCells = TRUE,
  removeZeroFeatures = TRUE,
  removeNonCanonical = TRUE,
  removeChrM = TRUE
)
}
\arguments{
\item{datamatrix}{}

\item{annot}{}

\item{removeZeroCells}{}

\item{removeZeroFeatures}{}
}
\value{

}
\description{
Title
}
