% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{DA_custom}
\alias{DA_custom}
\title{Differential Analysis Custom in 'One vs One' mode}
\usage{
DA_custom(
  affectation,
  by,
  counts,
  method,
  feature,
  block,
  ref,
  group,
  progress = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{affectation}{An annotation data.frame with cell_id and}

\item{by}{= A character specifying the column of the object containing
the groups of cells to compare.}

\item{counts}{Count matrix}

\item{method}{DA method : Wilcoxon or EdgeR}

\item{feature}{Feature tables}

\item{block}{Blocking feature}

\item{ref}{If de_type is custom, the reference to compare (data.frame), must
be a one-column data.frame with cell_clusters or sample_id as character in
rows}

\item{group}{If de_type is custom, the group to compare (data.frame), must
be a one-column data.frame with cell_clusters or sample_id as character in
rows}

\item{progress}{A shiny Progress instance to display progress bar.}

\item{BPPARAM}{BPPARAM object for multiprocessing. See
\link[BiocParallel]{bpparam} for more informations. Will take the default
BPPARAM set in your R session.}
}
\value{
A list of results, groups compared and references
}
\description{
Differential Analysis Custom in 'One vs One' mode
}
