% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{DA_one_vs_rest}
\alias{DA_one_vs_rest}
\title{Differential Analysis in 'One vs Rest' mode}
\usage{
DA_one_vs_rest(
  affectation,
  by,
  counts,
  method,
  feature,
  block,
  progress = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{affectation}{An annotation data.frame with cell_id and cell_cluster
columns}

\item{by}{= A character specifying the column of the object containing
the groups of cells to compare.}

\item{counts}{Count matrix}

\item{method}{DA method : Wilcoxon or EdgeR}

\item{feature}{Feature tables}

\item{block}{Blocking feature}

\item{progress}{A shiny Progress instance to display progress bar.}

\item{BPPARAM}{BPPARAM object for multiprocessing. See
\link[BiocParallel]{bpparam} for more informations. Will take the default
BPPARAM set in your R session.}
}
\value{
A list of results, groups compared and references
}
\description{
Differential Analysis in 'One vs Rest' mode
}
