% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{DA_pairwise}
\alias{DA_pairwise}
\title{Run differential analysis in Pairwise mode}
\usage{
DA_pairwise(
  affectation,
  by,
  counts,
  method,
  feature,
  block,
  progress = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{affectation}{An annotation data.frame with cell_cluster and cell_id
columns}

\item{by}{= A character specifying the column of the object containing
the groups of cells to compare.}

\item{counts}{Count matrix}

\item{method}{DA method, Wilcoxon or edgeR}

\item{feature}{Feature data.frame}

\item{block}{Blocking feature}

\item{progress}{A shiny Progress instance to display progress bar.}

\item{BPPARAM}{BPPARAM object for multiprocessing. See
\link[BiocParallel]{bpparam} for more informations. Will take the default
BPPARAM set in your R session.}
}
\value{
A list of results, groups compared and references
}
\description{
Run differential analysis in Pairwise mode
}
