% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{detect_samples}
\alias{detect_samples}
\title{Heuristic discovery of samples based on cell labels}
\usage{
detect_samples(barcodes, nb_samples = 1)
}
\arguments{
\item{barcodes}{}

\item{nb_samples}{}
}
\value{
character vector of sample names the same length as cell labels
}
\description{
Heuristic discovery of samples based on cell labels
}
