% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{differential_activation}
\alias{differential_activation}
\title{Find Differentialy Activated Features}
\usage{
differential_activation(
  scExp,
  group_by = c("cell_cluster", "sample_id")[1],
  progress = NULL
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing consclust with selected
number of cluster.}

\item{group_by}{Which grouping to run the marker enrichment ?}

\item{progress}{A shiny Progress instance to display progress bar.}

\item{prioritize_genes}{First filter by loci being close to genes ? E.g. for
differential analysis, it is more relevant to keep features close to genes}

\item{max_distanceToTSS}{If prioritize_genes is TRUE, the maximum distance to
consider a feature close to a gene.}

\item{BPPARAM}{BPPARAM object for multiprocessing. See
\link[BiocParallel]{bpparam} for more informations. Will take the default
BPPARAM set in your R session.}
}
\value{
Returns a dataframe of differential activation results.
}
\description{
Find Differentialy Activated Features
}
\examples{
data("scExp")
res = differential_activation(scExp, group_by = "cell_cluster")
}
