% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{differential_activation_scExp}
\alias{differential_activation_scExp}
\title{Find Differentialy Activated Features}
\usage{
differential_activation_scExp(
  scExp,
  group_by = c("cell_cluster", "sample_id"),
  prioritize_genes = nrow(scExp) > 20000,
  max_distanceToTSS = 1000,
  qval.th = 0.01,
  logFC.th = 1,
  min.percent = 0.01,
  progress = NULL,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing consclust with selected
number of cluster.}

\item{group_by}{Which grouping to run the marker enrichment ?}

\item{prioritize_genes}{First filter by loci being close to genes ? E.g. for
differential analysis, it is more relevant to keep features close to genes}

\item{max_distanceToTSS}{If prioritize_genes is TRUE, the maximum distance to
consider a feature close to a gene.}

\item{qval.th}{Adjusted p-value threshold. (0.01)}

\item{logFC.th}{Fold change threshold. (1)}

\item{min.percent}{Minimum fraction of cells having the feature active to
consider it as significantly differential. (0.01)}

\item{progress}{A shiny Progress instance to display progress bar.}

\item{BPPARAM}{BPPARAM object for multiprocessing. See
\link[BiocParallel]{bpparam} for more informations. Will take the default
BPPARAM set in your R session.}
}
\value{
Returns a SingleCellExperiment object containing a differential list.
}
\description{
Find Differentialy Activated Features
}
\examples{
data("scExp")
scExp_cf = differential_activation_scExp(scExp, group_by = "cell_cluster")
}
