% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{distPearson}
\alias{distPearson}
\title{distPearson}
\usage{
distPearson(m)
}
\arguments{
\item{m}{}
}
\value{
A dist object
}
\description{
distPearson
}
\examples{
m = matrix(runif(100),runif(100),ncol=100,nrow=100)
distPearson(m)
}
