% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{enrichmentTest}
\alias{enrichmentTest}
\title{enrichmentTest}
\usage{
enrichmentTest(gene.sets, mylist, possibleIds, sep = ";", silent = FALSE)
}
\arguments{
\item{gene.sets}{A list of reference gene sets}

\item{mylist}{A list of genes to test}

\item{possibleIds}{All existing genes}

\item{sep}{Separator used to collapse genes}

\item{silent}{Silent mode ?}
}
\value{
A dataframe with the gene sets and their enrichment p.value
}
\description{
enrichmentTest
}
