% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{exclude_features_scExp}
\alias{exclude_features_scExp}
\title{Remove specific features (CNA, repeats)}
\usage{
exclude_features_scExp(scExp, features_to_exclude, by = "region", verbose = T)
}
\arguments{
\item{scExp}{A SingleCellExperiment object.}

\item{features_to_exclude}{A data.frame containing features to exclude.}

\item{by}{Type of features. Either 'region' or 'feature_name'. If 'region',
 will look for genomic coordinates in columns 1-3 (chr,start,stop). 
If 'feature_name', will look for a genes in first column. ['region']}

\item{verbose}{[T]}
}
\value{
A SingleCellExperiment object without features to exclude.
}
\description{
Remove specific features (CNA, repeats)
}
