% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{fast_approx_cor_clustering}
\alias{fast_approx_cor_clustering}
\title{Fast approximated correlation hierarchical clustering}
\usage{
fast_approx_cor_clustering(
  pca,
  n_cell_per_metacell = 20,
  seed = 47,
  correlation = "pearson",
  hc_linkage = "ward.D"
)
}
\arguments{
\item{pca}{}

\item{n_cell_per_metacell}{}

\item{seed}{}

\item{correlation}{}

\item{hc_linkage}{}
}
\value{

}
\description{
Fast approximated correlation hierarchical clustering
}
