% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{fast_heatmap}
\alias{fast_heatmap}
\title{Fast approximated correlation hierarchical clustering}
\usage{
fast_heatmap(
  scExp_cf,
  n_cell_per_metacell = 15,
  cluster = NULL,
  seed = 47,
  correlation = "pearson",
  hc_linkage = "ward.D"
)
}
\arguments{
\item{correlation}{Type of correlation}

\item{hc_linkage}{Type of hierarchical clustering}
}
\value{
A list of clusters
}
\description{
Fast approximated correlation hierarchical clustering
}
\examples{
fast_approx_cor_clustering()
}
