% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{filter_scExp}
\alias{filter_scExp}
\title{Title}
\usage{
filter_scExp(
  scExp,
  min_cov_cell = 1600,
  quant_removal = 95,
  percentMin = 1,
  bin_min_count = 2,
  non_canonical = T,
  chrM = T,
  verbose = T
)
}
\arguments{
\item{scExp}{}

\item{min_cov_cell}{}

\item{quant_removal}{}

\item{percentMin}{}

\item{bin_min_count}{}
}
\value{

}
\description{
Title
}
