% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.annotToCol2}
\alias{geco.annotToCol2}
\title{geco.annotToCol2}
\usage{
geco.annotToCol2(
  annotS = NULL,
  annotT = NULL,
  missing = c("", NA),
  anotype = NULL,
  maxnumcateg = 2,
  categCol = NULL,
  quantitCol = NULL,
  plotLegend = T,
  plotLegendFile = NULL
)
}
\arguments{
\item{annotS}{}

\item{annotT}{}

\item{missing}{}

\item{anotype}{}

\item{maxnumcateg}{}

\item{categCol}{}

\item{quantitCol}{}

\item{plotLegend}{}

\item{plotLegendFile}{}
}
\value{

}
\description{
geco.annotToCol2
}
