% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.changeRange}
\alias{geco.changeRange}
\title{geco.changeRange}
\usage{
geco.changeRange(v, newmin = 1, newmax = 10)
}
\arguments{
\item{v}{}

\item{newmin}{}

\item{newmax}{}
}
\value{

}
\description{
geco.changeRange
}
