% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.enrichmentTest}
\alias{geco.enrichmentTest}
\title{geco.enrichmentTest}
\usage{
geco.enrichmentTest(gene.sets, mylist, possibleIds, sep = ";", silent = F)
}
\arguments{
\item{gene.sets}{}

\item{mylist}{}

\item{possibleIds}{}

\item{sep}{}

\item{silent}{}
}
\value{

}
\description{
geco.enrichmentTest
}
