% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.groupMat}
\alias{geco.groupMat}
\title{Title}
\usage{
geco.groupMat(mat = NA, margin = 1, groups = NA, method = "mean")
}
\arguments{
\item{mat}{}

\item{margin}{}

\item{groups}{}

\item{method}{}
}
\value{

}
\description{
Title
}
