% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.H1proportion}
\alias{geco.H1proportion}
\title{geco.H1proportion}
\usage{
geco.H1proportion(pv = NA, lambda = 0.5)
}
\arguments{
\item{pv}{}

\item{lambda}{}
}
\value{

}
\description{
geco.H1proportion
}
