% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.imageCol}
\alias{geco.imageCol}
\title{geco.imageCol}
\usage{
geco.imageCol(
  matcol = NULL,
  strat = NULL,
  xlab.cex = 0.5,
  ylab.cex = 0.5,
  drawLines = c("none", "h", "v", "b")[1],
  ...
)
}
\arguments{
\item{matcol}{}

\item{strat}{}

\item{xlab.cex}{}

\item{ylab.cex}{}

\item{drawLines}{}

\item{...}{}
}
\value{

}
\description{
geco.imageCol
}
