% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{generate_count_matrix}
\alias{generate_count_matrix}
\title{Generate count matrix}
\usage{
generate_count_matrix(cells, features, sparse, cell_names, feature_names)
}
\arguments{
\item{cells}{Number of cells}

\item{features}{Number of features}

\item{sparse}{Is matrix sparse ?}

\item{cell_names}{Cell names}

\item{feature_names}{Feature names}
}
\value{
A matrix or a sparse matrix
}
\description{
Generate count matrix
}
