% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{generate_feature_names}
\alias{generate_feature_names}
\title{Generate feature names}
\usage{
generate_feature_names(featureType, ref, features)
}
\arguments{
\item{featureType}{Type of feature}

\item{ref}{Reference genome}

\item{features}{Number of features to generate}
}
\value{
A character vector of feature names
}
\description{
Generate feature names
}
