% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_color_dataframe_from_input}
\alias{get_color_dataframe_from_input}
\title{Get color dataframe from shiny::colorInput}
\usage{
get_color_dataframe_from_input(
  input,
  levels_selected,
  color_by = c("sample_id", "total_counts"),
  input_id_prefix = "color_"
)
}
\arguments{
\item{input}{}

\item{levels_selected}{}

\item{color_by}{}

\item{input_id_prefix}{}
}
\value{

}
\description{
Get color dataframe from shiny::colorInput
}
