% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getExperimentNames}
\alias{getExperimentNames}
\title{Get experiment names from a SingleCellExperiment}
\usage{
getExperimentNames(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment with named mainExp and altExps.}
}
\value{
Character vector of unique experiment names
}
\description{
Get experiment names from a SingleCellExperiment
}
\examples{
data(scExp)
getExperimentNames(scExp)

}
