% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getMainExperiment}
\alias{getMainExperiment}
\title{Get Main experiment of a SingleCellExperiment}
\usage{
getMainExperiment(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment with named mainExp and altExps.}
}
\value{
The swapped SingleCellExperiment towards "main" experiment
}
\description{
Get Main experiment of a SingleCellExperiment
}
\examples{
getMainExperiment(scExp)

}
