% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{groupMat}
\alias{groupMat}
\title{groupMat}
\usage{
groupMat(mat = NA, margin = 1, groups = NA, method = "mean")
}
\arguments{
\item{mat}{A matrix}

\item{margin}{By row or columns ?}

\item{groups}{Groups}

\item{method}{Method to group}
}
\value{
A grouped matrix
}
\description{
groupMat
}
