% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{has_genomic_coordinates}
\alias{has_genomic_coordinates}
\title{Does SingleCellExperiment has genomic coordinates in features ?}
\usage{
has_genomic_coordinates(scExp)
}
\arguments{
\item{scExp}{}
}
\value{
TRUE or FALSE
}
\description{
Does SingleCellExperiment has genomic coordinates in features ?
}
\examples{

scExp = create_scExp(create_scDataset_raw()$mat,  create_scDataset_raw()$annot)
has_genomic_coordinates(scExp)
scExp_gene = create_scExp(create_scDataset_raw(featureType="gene")$mat,
  create_scDataset_raw(featureType="gene")$annot)
has_genomic_coordinates(scExp_gene)

}
