% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.cytoBand}
\alias{hg38.cytoBand}
\title{Data.frame of cytoBandlocation - hg38}
\format{
hg38.cytoBand - a data frame with 862 cytobands and 4 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the chromosome (bp) - integer}
\item{end}{End of the chromosome (bp) - integer}
\item{cytoBand}{Name of the cytoBand - character}
}
}
\usage{
data("hg38.cytoBand")
}
\value{
hg38.cytoBand - a data frame with 862 cytobands and 4 variables.
}
\description{
This data frame provides the location of each cytoBands of
Homo Sapiens genome build hg38.
}
\keyword{datasets}
