% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.MSIG.gs}
\alias{hg38.MSIG.gs}
\title{Molecular pathways & Genes from MSIG database - hg38}
\format{
hg38.MSIG.gs - a data frame with 9548 rows and 3 variables:
\describe{
\item{Gene.Set}{Name of gene sets}
\item{Class}{8 categories of pathways}
\item{Genes}{Gene symbols}
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/index.jsp}
}
\usage{
hg38.MSIG.gs
}
\description{
data.frame of pathways & genes that compose each pathway.
}
\keyword{datasets}
