% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{imageCol}
\alias{imageCol}
\title{imageCol}
\usage{
imageCol(
  matcol = NULL,
  strat = NULL,
  xlab.cex = 0.5,
  ylab.cex = 0.5,
  drawLines = c("none", "h", "v", "b")[1],
  ...
)
}
\arguments{
\item{matcol}{A matrix of colors}

\item{strat}{Strat}

\item{xlab.cex}{X label size}

\item{ylab.cex}{Y label size}

\item{drawLines}{Draw lines ?}

\item{...}{Additional parameteres}
}
\value{
A rectangular image
}
\description{
imageCol
}
