% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{import_count_input_files}
\alias{import_count_input_files}
\title{Import and count input files depending on their format}
\usage{
import_count_input_files(
  files_dir_list,
  file_type,
  which,
  ref,
  feature_file,
  barcode_file,
  matrix_file,
  verbose
)
}
\arguments{
\item{file_type}{Input file type}

\item{which}{A GRanges object of features}

\item{ref}{Reference genome}

\item{feature_file}{A file for containing unique feature genomic locations
in tab separated format (.bed / .txt / .tsv / .*.gz), e.g. chr, start and end}

\item{barcode_file}{A file containing unique barcode names}

\item{matrix_file}{A file containing indexes of non zero entries}

\item{verbose}{Print ?}

\item{files_dir}{A named list of directories containing the input files}
}
\value{
A list with a GRanges object of feature types (which), the
feature indexes data.frame containing non-zeroes entries in the count matrix
and the cell names
}
\description{
Import and count input files depending on their format
}
