% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch ChromSCape}
\usage{
launchApp(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Wether to launch browser or not}

\item{...}{Additional parameters passed to \code{\link[shiny]{runApp}}}
}
\value{
Launches the shiny application
}
\description{
Main function to launch ChromSCape in your favorite browser. You can pass
additional parameters that you would pass to shiny::runApp
(\code{\link[shiny]{runApp}})
}
\examples{
\dontrun{
launchApp()
}
}
