% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{load_MSIGdb}
\alias{load_MSIGdb}
\title{Load and format MSIGdb pathways using msigdbr package}
\usage{
load_MSIGdb(ref)
}
\arguments{
\item{ref}{Reference genome, either mm10 or hg38}
}
\value{
A list containing the GeneSet (list), GeneSetDf (data.frame) and
GenePool character vector of all possible genes
}
\description{
Load and format MSIGdb pathways using msigdbr package
}
\examples{
load_MSIGdb("hg38") 
load_MSIGdb("mm10") 
}
