% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{merge_MACS2_peaks}
\alias{merge_MACS2_peaks}
\title{Merge peak files from MACS2 peak caller}
\usage{
merge_MACS2_peaks(peak_file, peak_distance_to_merge, min_peak_size = 200, ref)
}
\arguments{
\item{peak_file}{A character specifying the path towards the peak file (BED
or bedGraph format)}

\item{peak_distance_to_merge}{Maximum distance to merge two peaks}

\item{min_peak_size}{An integer specifying the minimum size of peaks}

\item{ref}{Reference genome}
}
\value{
Peaks as GRanges
}
\description{
Merge peak files from MACS2 peak caller
}
