% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.chromosomes}
\alias{mm10.chromosomes}
\title{Data.frame of chromosome length - mm10}
\format{
mm10.chromosomes - a data frame with 24 chromosomes and 3 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the chromosome (bp) - integer}
\item{end}{End of the chromosome (bp) - integer}
}
}
\usage{
data("mm10.chromosomes")
}
\value{
mm10.chromosomes - a data frame with 24 chromosomes and 3 variables
}
\description{
This data frame provides the length of each "canonical" chromosomes of
Mus Musculus (Mouse) genome build mm10.
}
\keyword{datasets}
