% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.GeneTSS}
\alias{mm10.GeneTSS}
\title{Data.frame of gene TSS - mm10}
\format{
mm10.GeneTSS - a data frame with 27,916 genes and 5 variables:
\describe{
\item{chr}{Chromosome name - character}
\item{start}{Start of the gene (TSS) - integer}
\item{end}{End of the gene - integer}
\item{Gene}{Gene symbol - character}
\item{Strand}{Strand - character}
}
}
\usage{
data("mm10.GeneTSS")
}
\value{
mm10.GeneTSS - a data frame with 27,916 genes and 5 variables
}
\description{
This dataframe was extracted from Gencode v25 and report the Transcription
Start Site of each gene in the Mus Musculus genome build mm10 (Mouse).
}
\keyword{datasets}
