% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.MSIG.gs}
\alias{mm10.MSIG.gs}
\title{Molecular pathways & Genes from MSIG database - mm10}
\format{
mm10.MSIG.gs - a data frame with 9548 rows and 3 variables:
\describe{
  \item{Gene.Set}{Name of gene sets}
  \item{Class}{8 categories of pathways}
  \item{Genes}{Gene symbols}
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/index.jsp}
}
\usage{
mm10.MSIG.gs
}
\description{
data.frame of pathways & genes that compose each pathway.
}
\keyword{datasets}
