% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.MSIG.ls}
\alias{mm10.MSIG.ls}
\title{Molecular pathways & Genes from MSIG database - mm10}
\format{
mm10.MSIG.ls - a list with 9548 elements:
\describe{
  \item{names}{Name of gene sets}
  \item{values}{Gene symbols}
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/index.jsp}
}
\usage{
mm10.MSIG.ls
}
\description{
List of pathways & genes that compose each pathway.
}
\keyword{datasets}
