% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{num_cell_after_cor_filt_scExp}
\alias{num_cell_after_cor_filt_scExp}
\title{Number of cells before & after correlation filtering}
\usage{
num_cell_after_cor_filt_scExp(scExp, scExp_cf)
}
\arguments{
\item{scExp}{SingleCellExperiment object before correlation filtering.}

\item{scExp_cf}{SingleCellExperiment object atfer correlation filtering.}
}
\value{

}
\description{
Number of cells before & after correlation filtering
}
