% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{num_cell_in_cluster_scExp}
\alias{num_cell_in_cluster_scExp}
\title{Number of cells in each cluster}
\usage{
num_cell_in_cluster_scExp(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing chromatin groups.}
}
\value{
A formatted kable of cell assignation to each cluster.
}
\description{
Number of cells in each cluster
}
\examples{
data("scExp")
scExp_cf = correlation_and_hierarchical_clust_scExp(scExp)
scExp_cf = choose_cluster_scExp(scExp_cf,nclust=3,consensus=FALSE)
\dontrun{num_cell_in_cluster_scExp(scExp_cf)}

}
