% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{pca_irlba_for_sparseMatrix}
\alias{pca_irlba_for_sparseMatrix}
\title{Run sparse PCA using irlba SVD}
\usage{
pca_irlba_for_sparseMatrix(x, n_comp)
}
\arguments{
\item{n_comp}{}
}
\value{

}
\description{
Run sparse PCA using irlba SVD
}
