% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_cluster_consensus_scExp}
\alias{plot_cluster_consensus_scExp}
\title{Plot cluster consensus}
\usage{
plot_cluster_consensus_scExp(scExp)
}
\arguments{
\item{cname}{}
}
\value{

}
\description{
Plot cluster consensus score for each k as a bargraph.
}
