% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_coverage_BigWig}
\alias{plot_coverage_BigWig}
\title{Coverage plot using Sushi}
\usage{
plot_coverage_BigWig(
  coverages,
  label_color_list,
  peaks = NULL,
  chrom,
  start,
  end,
  ref = "hg38"
)
}
\arguments{
\item{coverages}{A list containing sample coverage as GenomicRanges}

\item{label_color_list}{List of colors, list names are labels}

\item{peaks}{A GRanges object containing peaks location to plot (optional)}

\item{chrom}{Chromosome}

\item{start}{Start}

\item{end}{End}

\item{ref}{Genomic Reference}
}
\value{
A coverage plot annotated with genes
}
\description{
Coverage plot using Sushi
}
\examples{
data(scExp)
plot_intra_correlation_scExp(scExp)

}
