% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_differential_H1_scExp}
\alias{plot_differential_H1_scExp}
\title{Differential H1 distribution plot}
\usage{
plot_differential_H1_scExp(scExp_cf, cell_cluster = "C1")
}
\arguments{
\item{scExp_cf}{}

\item{cell_cluster}{}
}
\value{

}
\description{
Differential H1 distribution plot
}
