% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_differential_volcano_scExp}
\alias{plot_differential_volcano_scExp}
\title{Volcano plot of differential features}
\usage{
plot_differential_volcano_scExp(
  scExp_cf,
  cell_cluster = "C1",
  cdiff.th = 1,
  qval.th = 0.01
)
}
\arguments{
\item{scExp_cf}{A SingleCellExperiment object}

\item{cell_cluster}{Which cluster to plot}

\item{cdiff.th}{Fold change threshold}

\item{qval.th}{Adjusted p.value threshold}
}
\value{
A volcano plot of differential analysis of a specific cluster
}
\description{
Volcano plot of differential features
}
\examples{
data("scExp")
plot_differential_volcano_scExp(scExp,"C1")
}
