% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_distribution_scExp}
\alias{plot_distribution_scExp}
\title{Plotting distribution of signal}
\usage{
plot_distribution_scExp(
  scExp,
  raw = T,
  log10 = F,
  pseudo_counts = 1,
  bins = 150
)
}
\arguments{
\item{scExp}{}

\item{raw}{}

\item{log10}{}

\item{pseudo_counts}{}

\item{bins}{}
}
\value{

}
\description{
Plotting distribution of signal
}
\examples{

}
