% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_heatmap_scExp}
\alias{plot_heatmap_scExp}
\title{Plot cell correlation heatmap with annotations}
\usage{
plot_heatmap_scExp(
  scExp,
  name_hc = "hc_cor",
  corColors = (grDevices::colorRampPalette(c("royalblue", "white", "indianred1")))(256),
  color_by = NULL
)
}
\arguments{
\item{scExp}{}

\item{name_hc}{}

\item{corColors}{}

\item{color_by}{}
}
\value{

}
\description{
Plot cell correlation heatmap with annotations
}
