% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_intra_correlation_scExp}
\alias{plot_intra_correlation_scExp}
\title{Violin plot of intra-correlation distribution}
\usage{
plot_intra_correlation_scExp(
  scExp_cf,
  by = c("sample_id", "cell_cluster")[1],
  jitter_by = NULL,
  downsample = 5000
)
}
\arguments{
\item{scExp_cf}{A SingleCellExperiment}

\item{by}{Color by sample_id or cell_cluster}

\item{jitter_by}{Add jitter points of another layer
(cell_cluster or sample_id)}

\item{downsample}{Downsample for plotting}
}
\value{
A violin plot of intra-correlation
}
\description{
Violin plot of intra-correlation distribution
}
\examples{
plot_intra_correlation_scExp(scExp)
}
