% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_reduced_dim_scExp}
\alias{plot_reduced_dim_scExp}
\title{Plot reduced dimensions (PCA, TSNE)}
\usage{
plot_reduced_dim_scExp(
  scExp,
  color_by = "sample_id",
  reduced_dim = c("PCA", "TSNE"),
  select_x = "PC1",
  select_y = "PC2",
  annot_label = "none"
)
}
\arguments{
\item{scExp}{}

\item{color_by}{}

\item{reduced_dim}{}

\item{select_x}{}

\item{select_y}{}

\item{annot_label}{}
}
\value{

}
\description{
Plot reduced dimensions (PCA, TSNE)
}
