% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{preprocess_feature_size_only}
\alias{preprocess_feature_size_only}
\title{Preprocess scExp - size only}
\usage{
preprocess_feature_size_only(scExp)
}
\arguments{
\item{scExp}{}
}
\value{
A SingleCellExperiment object.
}
\description{
Preprocess scExp - size only
}
