% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{raw_counts_to_feature_count_files}
\alias{raw_counts_to_feature_count_files}
\title{Create a sparse count matrix from various format of input data.}
\usage{
raw_counts_to_feature_count_files(
  files_dir,
  file_type = c("BAM", "BED", "Index_Peak_Barcode"),
  peak_file = NULL,
  n_bins = NULL,
  bin_width = NULL,
  geneTSS = NULL,
  aroundTSS = 2500,
  verbose = TRUE,
  ref = "hg38"
)
}
\arguments{
\item{file_type}{}

\item{ref}{reference genome to use \link{hg38}}
}
\value{
A sparse matrix of features x cells
}
\description{
This function takes three different type of single-cell input:
\itemize{
\item Single cell BAM files (sorted)
\item Single cell BED files (gzipped)
\item A combination of an index file, a peak file and cell barcode file
(The index file is composed of three column: index i, index j and value x for the
non zeroes entries in the sparse matrix. @seealso Matrix::sparseMatrix)
}

This functions re-counts signal on either fixed genomic bins, a set of
user-defined peaks or around the TSS of genes.
}
