% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{raw_counts_to_sparse_matrix}
\alias{raw_counts_to_sparse_matrix}
\title{Create a sparse count matrix from various format of input data.}
\usage{
raw_counts_to_sparse_matrix(
  files_dir_list,
  file_type = c("scBED", "scBAM"),
  peak_file = NULL,
  n_bins = NULL,
  bin_width = NULL,
  geneTSS = NULL,
  aroundTSS = 2500,
  verbose = TRUE,
  ref = c("hg38", "mm10")[1],
  progress = NULL
)
}
\arguments{
\item{files_dir_list}{A named character vector of directories containing
the files. The names correspond to sample names.}

\item{file_type}{Input file(s) type(s) ('scBED','scBAM','SparseMatrix')}

\item{peak_file}{A file containing genomic location of peaks (NULL)}

\item{n_bins}{The number of bins to tile the genome (NULL)}

\item{bin_width}{The size of bins to tile the genome (NULL)}

\item{geneTSS}{Use geneTSS regions for annotation ? (NULL)}

\item{aroundTSS}{Space up and downstream of TSS to use (2500)}

\item{verbose}{Verbose (TRUE)}

\item{ref}{reference genome to use (hg38)}

\item{progress}{Progress object for Shiny}
}
\value{
A sparse matrix of features x cells
}
\description{
This function takes three different type of single-cell input: - Single cell
BAM files (sorted) - Single cell BED files (gzipped) - A combination of an
index file, a peak file and cell barcode file (The index file is composed of
three column: index i, index j and value x for the non zeroes entries in the
sparse matrix.)
}
\details{
This functions re-counts signal on either fixed genomic bins, a set of
user-defined peaks or around the TSS of genes.
}
