% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{read_count_mat_with_separated_chr_start_end}
\alias{read_count_mat_with_separated_chr_start_end}
\title{Read a count matrix with three first columns (chr,start,end)}
\usage{
read_count_mat_with_separated_chr_start_end(
  path_to_matrix,
  format_test,
  separator
)
}
\arguments{
\item{path_to_matrix}{Path to the count matrix}

\item{format_test}{Sample of the read.table}

\item{separator}{Separator character}
}
\value{
A sparseMatrix with rownames in the form "chr1:1222-55555"
}
\description{
Read a count matrix with three first columns (chr,start,end)
}
