% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{rebin_helper}
\alias{rebin_helper}
\title{Rebin Helper for rebin_matrix function}
\usage{
rebin_helper(mat_df)
}
\arguments{
\item{mat_df}{A data.frame corresponding to sparse matrix indexes & values.}
}
\value{
a data.frame grouped mean-summarised  by col and new_row
}
\description{
Rebin Helper for rebin_matrix function
}
