% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{reduce_dims_scExp}
\alias{reduce_dims_scExp}
\title{Reduce dimensions (PCA, TSNE, UMAP)}
\usage{
reduce_dims_scExp(
  scExp,
  dimension_reductions = c("PCA", "TSNE"),
  n = 50,
  batch_correction = F,
  batch_list = NULL,
  verbose = T
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object.}

\item{dimension_reductions}{A character vector of methods to apply. [c("PCA","TSNE")]}

\item{n}{Numbers of dimensions to keep for PCA. [50]}

\item{verbose}{}
}
\value{
A SingleCellExperiment object containing feature spaces. See ?reduceDims().
}
\description{
Reduce dimensions (PCA, TSNE, UMAP)
}
