% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{remove_chr_M_fun}
\alias{remove_chr_M_fun}
\title{Remove chromosome M from scExprownames}
\usage{
remove_chr_M_fun(scExp, verbose)
}
\arguments{
\item{scExp}{A SingleCellExperiment}

\item{verbose}{Print ?}
}
\value{
A SingleCellExperiment without chromosome M (mitochondrial chr)
}
\description{
Remove chromosome M from scExprownames
}
