% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{remove_non_canonical_fun}
\alias{remove_non_canonical_fun}
\title{Remove non canonical chromosomes from scExp}
\usage{
remove_non_canonical_fun(scExp, verbose)
}
\arguments{
\item{scExp}{A SingleCellExperiment}

\item{verbose}{Print ?}
}
\value{
A SingleCellExperiment without non canonical chromosomes
(random,unknown, contigs etc...)
}
\description{
Remove non canonical chromosomes from scExp
}
