% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{results_enrichmentTest}
\alias{results_enrichmentTest}
\title{Resutls of hypergeometric gene set enrichment test}
\usage{
results_enrichmentTest(
  differentialGenes,
  enrichment_qval,
  GeneSets,
  GeneSetsDf,
  GenePool
)
}
\arguments{
\item{enrichment_qval}{Adusted p-value threshold above which a pathway is
considered significative}

\item{GeneSets}{List of pathways}

\item{GeneSetsDf}{Data.frame of pathways}

\item{GenePool}{Pool of possible genes for testing}

\item{significG}{Genes significantly differential}
}
\value{
A data.frame with pathways passing q.value threshold
}
\description{
Run hypergeometric enrichment test and combine significant pathways into
a data.frame
}
