% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{retrieve_top_bot_features_pca}
\alias{retrieve_top_bot_features_pca}
\title{Retrieve Top and Bot most contributing features of PCA}
\usage{
retrieve_top_bot_features_pca(
  pca,
  counts,
  component,
  n_top_bot,
  absolute = FALSE
)
}
\arguments{
\item{pca}{A matrix/data.frame of rotated data}

\item{counts}{the normalized counts used for PCA}

\item{component}{the componenent of interest}

\item{n_top_bot}{the number of top & bot features to take}

\item{absolute}{If TRUE, return the top features in absolute values instead.}
}
\value{
a data.frame of top bot contributing features in PCA
}
\description{
Retrieve Top and Bot most contributing features of PCA
}
