% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{run_tsne_scExp}
\alias{run_tsne_scExp}
\title{Run tsne on single cell experiment}
\usage{
run_tsne_scExp(scExp, verbose = FALSE)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}

\item{verbose}{Print ?}
}
\value{
A colored kable with the number of cells per sample for display
}
\description{
Run tsne on single cell experiment
}
