% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scExp_cf}
\alias{scExp_cf}
\title{A randomly generated SingleCellExperiment normalized and with reduced
dimensions, processed until GSA (hg38)}
\format{
scExp_cf - a SingleCellExperiment with 300 cells and 600 features
(genomic bins) in hg38 and processed until GSA
: \describe{ \item{chr}{A SingleCellExperiment} }
}
\usage{
scExp_cf
}
\description{
A randomly generated SingleCellExperiment normalized and with reduced
dimensions, processed until GSA (hg38)
}
\keyword{datasets}
