% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{smoothBin}
\alias{smoothBin}
\title{Smooth a vector of values with nb_bins left and righ values}
\usage{
smoothBin(bin_score, nb_bins = 10)
}
\arguments{
\item{bin_score}{A numeric vector of values to be smoothed}

\item{nb_bins}{Number of values to take left and right

@importFrom BiocParallel bpvec}
}
\value{
A smooth vector of the same size
}
\description{
Smooth a vector of values with nb_bins left and righ values
}
