% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{subsample_scExp}
\alias{subsample_scExp}
\title{Subsample scExp}
\usage{
subsample_scExp(scExp, n_cells = 500)
}
\arguments{
\item{scExp}{A SingleCellExperiment}

\item{n_cells}{An integer number of cells to subsample for each sample [500]}
}
\value{

}
\description{
Randomly sample x cells from each sample in a SingleCellExperiment to return
a subsampled SingleCellExperiment with all samples having maximum n cells. If
n is higher than the number of cell in a sample, this sample will not be
subsampled.
}
