% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{summary_DA}
\alias{summary_DA}
\title{Summary of the differential analysis}
\usage{
summary_DA(scExp, qval.th = 0.01, logFC.th = 1, min.percent = 0.01)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing consclust with selected
number of cluster.}

\item{qval.th}{Adjusted p-value threshold. (0.01)}

\item{logFC.th}{Fold change threshold. (1)}

\item{min.percent}{Minimum fraction of cells having the feature active to
consider it as significantly differential. (0.01)}
}
\value{
A table summary of the differential analysis
}
\description{
Summary of the differential analysis
}
\examples{
data('scExp')
summary_DA(scExp)
}
