% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{swapAltExp_sameColData}
\alias{swapAltExp_sameColData}
\title{Swap main & alternative Experiments, with fixed colData}
\usage{
swapAltExp_sameColData(scExp, alt)
}
\arguments{
\item{scExp}{A SingleCellExperiment}

\item{alt}{Name of the alternative experiment}
}
\value{
A swapped SingleCellExperiment with the exact same colData.
}
\description{
Swap main & alternative Experiments, with fixed colData
}
\examples{
swapAltExp_sameColData(scExp, "peaks")

}
