% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{warning_DA}
\alias{warning_DA}
\title{Warning for differential_analysis_scExp}
\usage{
warning_DA(scExp, de_type, method, qval.th, cdiff.th, block)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing consclust with selected
number of cluster.}

\item{de_type}{Type of comparisons. Either 'one_vs_rest', to compare each
cluster against all others, or 'pairwise' to make 1 to 1 comparisons.
('one_vs_rest')}

\item{method}{Wilcoxon or edgerGLM}

\item{qval.th}{Adjusted p-value threshold. (0.01)}

\item{cdiff.th}{Fold change threshold. (1)}

\item{block}{Use batches as blocking factors ?}
}
\value{
Warnings or Errors if the input are not correct
}
\description{
Warning for differential_analysis_scExp
}
