% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_filtering_clustering.R
\name{warning_filter_correlated_cell_scExp}
\alias{warning_filter_correlated_cell_scExp}
\title{warning_filter_correlated_cell_scExp}
\usage{
warning_filter_correlated_cell_scExp(
  scExp,
  random_iter,
  corr_threshold,
  percent_correlation,
  run_tsne,
  verbose
)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing 'Cor', a correlation
matrix, in reducedDims.}

\item{random_iter}{Number of random matrices to create to calculate random
correlation scores. (50)}

\item{corr_threshold}{Quantile of random correlation score above which a cell
is considered to be 'correlated' with another cell. (99)}

\item{percent_correlation}{Percentage of the cells that any cell must be
'correlated' to in order to not be filtered. (1)}

\item{run_tsne}{Re-run tsne ? (FALSE)}

\item{verbose}{(TRUE)}
}
\description{
warning_filter_correlated_cell_scExp
}
