% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{warning_raw_counts_to_feature_count_files}
\alias{warning_raw_counts_to_feature_count_files}
\title{Warning for _raw_counts_to_feature_count_files}
\usage{
warning_raw_counts_to_feature_count_files(
  files_dir,
  file_type = c("BAM", "BED", "Index_Peak_Barcode"),
  peak_file = NULL,
  n_bins = NULL,
  bin_width = NULL,
  geneTSS = NULL,
  aroundTSS = 2500,
  verbose = TRUE,
  ref = "hg38"
)
}
\arguments{
\item{files_dir}{The directory containing the files}

\item{file_type}{Input file(s) type(s) ('BAM')}

\item{peak_file}{A file containing genomic location of peaks (NULL)}

\item{n_bins}{The number of bins to tile the genome (NULL)}

\item{bin_width}{The size of bins to tile the genome (NULL)}

\item{geneTSS}{Use geneTSS regions for annotation ? (NULL)}

\item{aroundTSS}{Space up and downstream of TSS to use (2500)}

\item{verbose}{Verbose (TRUE)}

\item{ref}{reference genome to use (hg38)}
}
\value{
Error or warnings if the input are not correct
}
\description{
Warning for _raw_counts_to_feature_count_files
}
