% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{annotToCol2}
\alias{annotToCol2}
\title{annotToCol2}
\usage{
annotToCol2(
  annotS = NULL,
  annotT = NULL,
  missing = c("", NA),
  anotype = NULL,
  maxnumcateg = 2,
  categCol = NULL,
  quantitCol = NULL,
  plotLegend = TRUE,
  plotLegendFile = NULL
)
}
\arguments{
\item{annotS}{A color matrix}

\item{annotT}{A color matrix}

\item{missing}{Convert missing to NA}

\item{anotype}{Annotation type}

\item{maxnumcateg}{Maximum number of categories}

\item{categCol}{Categorical columns}

\item{quantitCol}{Quantitative columns}

\item{plotLegend}{Plot legend ?}

\item{plotLegendFile}{Which file to plot legend ?}
}
\value{
A matrix of continuous or discrete colors
}
\description{
annotToCol2
}
\examples{
data("scExp")
annotToCol2(SingleCellExperiment::colData(scExp), plotLegend = FALSE)

}
