% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{bams_to_matrix_indexes}
\alias{bams_to_matrix_indexes}
\title{Count bam files on interval to create count indexes}
\usage{
bams_to_matrix_indexes(files_dir, which)
}
\value{
feature_indexes
}
\description{
Count bam files on interval to create count indexes
}
