% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{bams_to_matrix_indexes}
\alias{bams_to_matrix_indexes}
\title{Count bam files on interval to create count indexes}
\usage{
bams_to_matrix_indexes(files_dir, which)
}
\arguments{
\item{files_dir}{}

\item{which}{}
}
\value{
A list containing a "feature index" data.frame and a
count vector for non 0 entries, both used to form the sparse matrix
}
\description{
Count bam files on interval to create count indexes
}
