% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ce11.chromosomes}
\alias{ce11.chromosomes}
\title{Data.frame of chromosome length - ce11}
\format{
ce11.chromosomes - a data frame with 7 chromosomes and 3 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the chromosome (bp) - integer}
\item{end}{End of the chromosome (bp) - integer}
}
}
\usage{
data("ce11.chromosomes")
}
\description{
This data frame provides the length of each "canonical" chromosomes of
Homo Sapiens genome build ce11.
}
\keyword{datasets}
