% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{changeRange}
\alias{changeRange}
\title{changeRange}
\usage{
changeRange(v, newmin = 1, newmax = 10)
}
\arguments{
\item{v}{}

\item{newmin}{}

\item{newmax}{}
}
\value{

}
\description{
changeRange
}
