% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CheA3_TF_nTargets}
\alias{CheA3_TF_nTargets}
\title{A data.frame with the number of targets of each TF in ChEA3}
\format{
CheA3_TF_nTargets - a data.frame with 1632 rows (unique TFs) and
2 columns
}
\usage{
data("CheA3_TF_nTargets")
}
\description{
This data.frame was obtained by downloading datasets from ChEA3 database
(https://maayanlab.cloud/chea3/) and merging targets for :
\itemize{
\item ARCHS4_Coexpression
\item ENCODE_ChIP-seq
\item Enrichr_Queries
\item GTEx_Coexpression
\item Literature_ChIP-seq
\item ReMap_ChIP-seq
}
}
\examples{
data("CheA3_TF_nTargets")
head(CheA3_TF_nTargets)

}
\references{
Keenan AB, Torre D, Lachmann A, Leong AK, Wojciechowicz M, Utti V,
Jagodnik K, Kropiwnicki E, Wang Z, Ma'ayan A (2019)
ChEA3: transcription factor enrichment analysis by orthogonal omics integration.
Nucleic Acids Research. doi: 10.1093/nar/gkz446

The data.frame is composed of two columns:
\itemize{
\item TF column containing the TF gene names (human)
\item nTargets_TF containing the number of targets for this TF in the combined
database.
}
}
\keyword{datasets}
