% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{CompareWilcox}
\alias{CompareWilcox}
\title{CompareWilcox}
\usage{
CompareWilcox(
  dataMat = NULL,
  annot = NULL,
  ref = NULL,
  groups = NULL,
  featureTab = NULL,
  block = NULL
)
}
\arguments{
\item{dataMat}{}

\item{annot}{}

\item{ref}{}

\item{groups}{}

\item{featureTab}{}

\item{block}{}
}
\value{

}
\description{
CompareWilcox
}
