% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{concatenate_scBed_into_clusters}
\alias{concatenate_scBed_into_clusters}
\title{Concatenate single-cell BED into clusters}
\usage{
concatenate_scBed_into_clusters(affectation, files_list, odir)
}
\arguments{
\item{affectation}{Annotation data.frame containing cluster information}

\item{files_list}{Named list of scBED file paths to concatenate. List Names
must match affectation$sample_id and basenames must match
affectation$barcode.}

\item{odir}{Output directory to write concatenate pseudo-bulk BEDs.}
}
\value{
Merge single-cell BED files into cluster BED files. Ungzip file if
BED is gzipped.
}
\description{
Concatenate single-cell BED into clusters
}
