% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{create_scDataset_raw}
\alias{create_scDataset_raw}
\title{Create a simulated single cell datamatrix & cell annotation}
\usage{
create_scDataset_raw(
  cells = 300,
  features = 600,
  featureType = c("window", "peak", "gene"),
  sparse = T,
  nsamp = 4,
  ref = "hg38",
  batch_id = rep(1, nsamp)
)
}
\arguments{
\item{batch_id}{}
}
\value{

}
\description{
Create a simulated single cell datamatrix & cell annotation
}
