% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{DA_custom}
\alias{DA_custom}
\title{Differential Analysis in 'One vs Rest' mode}
\usage{
DA_custom(
  affectation,
  counts,
  method,
  feature,
  block,
  ref,
  group,
  progress = NULL
)
}
\arguments{
\item{affectation}{An annotation data.frame with cell_id and cell_cluster
columns}

\item{counts}{Count matrix}

\item{method}{DA method : Wilcoxon or EdgeR}

\item{feature}{Feature tables}

\item{block}{Blocking feature}

\item{progress}{A shiny Progress instance to display progress bar.}

\item{nclust}{Number of clusters}
}
\value{
A list of results, groups compared and references
}
\description{
Differential Analysis in 'One vs Rest' mode
}
