% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{DA_pairwise}
\alias{DA_pairwise}
\title{Run differential analysis in Pairwise mode}
\usage{
DA_pairwise(affectation, nclust, counts, method, feature, block)
}
\arguments{
\item{affectation}{An annotation data.frame with cell_cluster and cell_id
columns}

\item{nclust}{Number of clusters}

\item{counts}{Count matrix}

\item{method}{DA method, Wilcoxon or edgeR}

\item{feature}{Feature data.frame}

\item{block}{Blocking feature}
}
\value{
A list of results, groups compared and references
}
\description{
Run differential analysis in Pairwise mode
}
