% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{define_feature}
\alias{define_feature}
\title{Define the features on which reads will be counted}
\usage{
define_feature(ref, peak_file, n_bins, bin_width, geneTSS, aroundTSS)
}
\arguments{
\item{ref}{Reference genome}

\item{peak_file}{A bed file if counting on peaks}

\item{n_bins}{A number of bins if divinding genome into fixed number of bins}

\item{bin_width}{A number of bins if divinding genome into fixed width bins}

\item{geneTSS}{A logical indicating if feature should be counted around genes
TSS instead}

\item{aroundTSS}{Region to take in account around genes TSS}
}
\value{
A GRanges object
}
\description{
Define the features on which reads will be counted
}
