% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{enrichmentTest}
\alias{enrichmentTest}
\title{enrichmentTest}
\usage{
enrichmentTest(gene.sets, mylist, possibleIds, sep = ";", silent = F)
}
\arguments{
\item{gene.sets}{}

\item{mylist}{}

\item{possibleIds}{}

\item{sep}{}

\item{silent}{}
}
\value{

}
\description{
enrichmentTest
}
