% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_functions.R
\name{geco.CompareWilcox}
\alias{geco.CompareWilcox}
\title{geco.CompareWilcox}
\usage{
geco.CompareWilcox(
  dataMat = NULL,
  annot = NULL,
  ref = NULL,
  groups = NULL,
  featureTab = NULL,
  block = NULL
)
}
\arguments{
\item{dataMat}{}

\item{annot}{}

\item{ref}{}

\item{groups}{}

\item{featureTab}{}

\item{block}{}
}
\value{

}
\description{
geco.CompareWilcox
}
