% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_color_dataframe_from_input}
\alias{get_color_dataframe_from_input}
\title{Get color dataframe from shiny::colorInput}
\usage{
get_color_dataframe_from_input(
  input,
  levels_selected,
  color_by = c("sample_id", "total_counts"),
  input_id_prefix = "color_"
)
}
\arguments{
\item{input}{Shiny input object}

\item{levels_selected}{Names of the features}

\item{color_by}{Which feature color to retrieve}

\item{input_id_prefix}{Prefix in front of the feature names}
}
\value{
A data.frame with the feature levels and the colors of each level of
this feature.
}
\description{
Get color dataframe from shiny::colorInput
}
