% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{get_genomic_coordinates}
\alias{get_genomic_coordinates}
\title{Get SingleCellExperiment's genomic coordinates}
\usage{
get_genomic_coordinates(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment object.}
}
\value{
A GRanges object of genomic coordinates.
}
\description{
Get SingleCellExperiment's genomic coordinates
}
