% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{groupMat}
\alias{groupMat}
\title{groupMat}
\usage{
groupMat(mat = NA, margin = 1, groups = NA, method = "mean")
}
\arguments{
\item{mat}{}

\item{margin}{}

\item{groups}{}

\item{method}{}
}
\value{
A grouped matrix
}
\description{
groupMat
}
