% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{H1proportion}
\alias{H1proportion}
\title{H1proportion}
\usage{
H1proportion(pv = NA, lambda = 0.5)
}
\arguments{
\item{pv}{}

\item{lambda}{}
}
\value{
H1 proportion value
}
\description{
H1proportion
}
\examples{
p.value = runif(100)
H1proportion(p.value) = changeRange(mat,0,1)

}
