% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{has_genomic_coordinates}
\alias{has_genomic_coordinates}
\title{Does SingleCellExperiment has genomic coordinates in features ?}
\usage{
has_genomic_coordinates(scExp)
}
\arguments{
\item{scExp}{}
}
\value{
TRUE or FALSE
}
\description{
Does SingleCellExperiment has genomic coordinates in features ?
}
