% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{hclustAnnotHeatmapPlot}
\alias{hclustAnnotHeatmapPlot}
\title{hclustAnnotHeatmapPlot}
\usage{
hclustAnnotHeatmapPlot(
  x = NULL,
  hc = NULL,
  hmColors = NULL,
  anocol = NULL,
  xpos = c(0.1, 0.9, 0.114, 0.885),
  ypos = c(0.1, 0.5, 0.5, 0.6, 0.62, 0.95),
  dendro.cex = 1,
  xlab.cex = 0.8,
  hmRowNames = FALSE,
  hmRowNames.cex = 0.5
)
}
\arguments{
\item{x}{}

\item{hc}{}

\item{hmColors}{}

\item{anocol}{}

\item{xpos}{}

\item{ypos}{}

\item{dendro.cex}{}

\item{xlab.cex}{}

\item{hmRowNames}{}

\item{hmRowNames.cex}{}
}
\value{
A heatmap
}
\description{
hclustAnnotHeatmapPlot
}
