% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.chromosomes}
\alias{hg38.chromosomes}
\title{Data.frame of chromosome length - hg38}
\format{
hg38.chromosomes - a data frame with 24 rows and 3 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the chromosome (bp) - integer}
\item{end}{End of the chromosome (bp) - integer}
}
}
\usage{
data("hg38.chromosomes")
}
\description{
This data frame provides the length of each "canonical" chromosomes of
Homo Sapiens genome build hg38.
}
\keyword{datasets}
