% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.chromosomes}
\alias{hg38.chromosomes}
\title{Data.frame of chromosome length - hg38}
\format{
hg38.chromosomes - a data frame with 24 rows and 3 variables:
\describe{
\item{chr}{Chromosome name}
\item{star}{Start}
\item{end}{End}
}
}
\usage{
hg38.chromosomes
}
\description{
Data.frame of chromosome length - hg38
}
\keyword{datasets}
