% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.GeneTSS}
\alias{hg38.GeneTSS}
\title{Data.frame of gene TSS - hg38}
\format{
hg38.GeneTSS - a data frame with 24 rows and 3 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the gene (TSS) - integer}
\item{end}{End of the gene - integer}
\item{gene}{Gene symbol - character}
}
}
\usage{
data("hg38.GeneTSS")
}
\description{
This dataframe was extracted from Gencode v25 and report the Transcription
Start Site of each gene in the Homo Sapiens genome build hg38.
}
\keyword{datasets}
