% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38.GeneTSS}
\alias{hg38.GeneTSS}
\title{Data.frame of gene TSS - hg38}
\format{
hg38.GeneTSS - a data frame with 24 rows and 3 variables:
\describe{
  \item{chr}{Chromosome name}
  \item{start}{TSS start}
  \item{end}{TSS end}
  \item{gene}{Gene symbol}
}
}
\usage{
hg38.GeneTSS
}
\description{
Data.frame of gene TSS - hg38
}
\keyword{datasets}
