% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{import_count_input_files}
\alias{import_count_input_files}
\title{Import and count input files depending on their format}
\usage{
import_count_input_files(
  files_dir,
  file_type,
  which,
  ref,
  peak_file,
  peak_file_2,
  barcode_file,
  n_bins,
  bin_width,
  geneTSS,
  index_file,
  verbose
)
}
\arguments{
\item{files_dir}{Path to the input files}

\item{file_type}{Input file type}

\item{which}{A GRanges object of features}

\item{ref}{Reference genome}

\item{peak_file}{A bed file}

\item{peak_file_2}{A bed file for peak annotation}

\item{barcode_file}{A file containing barcode names}

\item{n_bins}{Number of bins}

\item{bin_width}{Bin width}

\item{geneTSS}{Gene TSS annotation}

\item{index_file}{A file containing indexes of non zero entries}

\item{verbose}{Print ?}
}
\value{
A list with a GRanges object of feature types (which), the
feature indexes data.frame containing non-zeroes entries in the count matrix
and the cell names
}
\description{
Import and count input files depending on their format
}
