% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{import_count_input_files}
\alias{import_count_input_files}
\title{Import and count input files depending on their format}
\usage{
import_count_input_files(files_dir_list, file_type, which, ref, verbose)
}
\arguments{
\item{file_type}{Input file type}

\item{which}{A GRanges object of features}

\item{ref}{Reference genome}

\item{verbose}{Print ?}

\item{files_dir}{A named list of directories containing the input files}
}
\value{
A list with the
feature indexes data.frame containing non-zeroes entries in the count matrix
and the cell names
}
\description{
Import and count input files depending on their format
}
