% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{import_count_input_files}
\alias{import_count_input_files}
\title{Import and count input files depending on their format}
\usage{
import_count_input_files(
  files_dir_list,
  file_type,
  which,
  ref,
  verbose,
  progress,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{file_type}{Input file type}

\item{ref}{Reference genome}

\item{verbose}{Print ?}

\item{BPPARAM}{BPPARAM object for multiprocessing. See
\link[BiocParallel]{bpparam} for more informations. Will take the default
BPPARAM set in your R session.}

\item{files_dir}{A named list of directories containing the input files}
}
\value{
A list with the
feature indexes data.frame containing non-zeroes entries in the count matrix
and the cell names
}
\description{
Import and count input files depending on their format
}
