% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{index_peaks_barcodes_to_matrix_indexes}
\alias{index_peaks_barcodes_to_matrix_indexes}
\title{Read index-peaks-barcodes trio files on interval to create count indexes}
\usage{
index_peaks_barcodes_to_matrix_indexes(
  feature_file,
  matrix_file,
  barcode_file,
  binarize = FALSE
)
}
\arguments{
\item{feature_file}{A file containing the features genomic locations}

\item{matrix_file}{A file containing the indexes of non-zeroes values and
their value (respectively i,j,x,see sparseMatrix)}

\item{barcode_file}{A file containing the barcode ids}

\item{binarize}{Binarize matrix ?}
}
\value{
A list containing a "feature index" data.frame, name_cells, and a
region GenomicRange object used to form the sparse matrix
}
\description{
Read index-peaks-barcodes trio files on interval to create count indexes
}
