% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{index_peaks_barcodes_to_matrix_indexes}
\alias{index_peaks_barcodes_to_matrix_indexes}
\title{Read index-peaks-barcodes trio files on interval to create count indexes}
\usage{
index_peaks_barcodes_to_matrix_indexes(
  peak_file,
  index_file,
  name_cells,
  binarize = FALSE,
  ref = "hg38"
)
}
\arguments{
\item{peak_file}{}

\item{index_file}{}

\item{name_cells}{}

\item{binarize}{}

\item{ref}{}
}
\value{
A list containing a "feature index" data.frame and a
region GenomicRange object both used to form the sparse matrix
}
\description{
Read index-peaks-barcodes trio files on interval to create count indexes
}
