% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{merge_MACS2_peaks}
\alias{merge_MACS2_peaks}
\title{Merge peak files from MACS2 peak caller}
\usage{
merge_MACS2_peaks(odir, class, peak_distance_to_merge, ref)
}
\arguments{
\item{odir}{Output directory}

\item{class}{Cell cluster}

\item{peak_distance_to_merge}{Maximum distance to merge two peaks}

\item{ref}{Reference genome}
}
\value{
A list of merged peaks
}
\description{
Merge peak files from MACS2 peak caller
}
