% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.chromosomes}
\alias{mm10.chromosomes}
\title{Data.frame of chromosome length - mm10}
\format{
mm10.chromosomes - a data frame with 24 rows and 3 variables:
\describe{
  \item{chr}{Chromosome name}
  \item{star}{Start}
  \item{end}{End}
}
}
\usage{
mm10.chromosomes
}
\description{
Data.frame of chromosome length - mm10
}
\keyword{datasets}
