% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.cytoBand}
\alias{mm10.cytoBand}
\title{Data.frame of cytoBandlocation - mm10}
\format{
mm10.cytoBand - a data frame with 862 rows and 4 variables:
\describe{
\item{chr}{Chromosome - character}
\item{start}{Start of the chromosome (bp) - integer}
\item{end}{End of the chromosome (bp) - integer}
\item{cytoBand}{Name of the cytoBand - character}
}
}
\usage{
data("mm10.cytoBand")
}
\description{
This data frame provides the location of each cytoBands of
Homo Sapiens genome build mm10.
}
\keyword{datasets}
