% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10.GeneTSS}
\alias{mm10.GeneTSS}
\title{Data.frame of gene TSS - mm10}
\format{
mm10.GeneTSS - a data frame with 24 rows and 3 variables:
\describe{
  \item{chr}{Chromosome name}
  \item{start}{TSS start}
  \item{end}{TSS end}
  \item{gene}{Gene symbol}
}
}
\usage{
mm10.GeneTSS
}
\description{
Data.frame of gene TSS - mm10
}
\keyword{datasets}
