% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{num_cell_scExp}
\alias{num_cell_scExp}
\title{Table of cells}
\usage{
num_cell_scExp(annot)
}
\arguments{
\item{scExp}{A SingleCellExperiment object.}
}
\value{
A formatted kable in HTML.
}
\description{
Table of cells
}
