% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{pca_irlba_for_sparseMatrix}
\alias{pca_irlba_for_sparseMatrix}
\title{Run sparse PCA using irlba SVD}
\usage{
pca_irlba_for_sparseMatrix(x, n_comp)
}
\arguments{
\item{n_comp}{}
}
\value{
The rotated data, e.g. the cells x PC column in case of sc data.
}
\description{
Run sparse PCA using irlba SVD
}
\examples{

pca_irlba_for_sparseMatrix(matrix(runif(100),nrow = 250,ncol=250),15)
pca_irlba_for_sparseMatrix(as(matrix(runif(100),nrow = 250,ncol=250),"dgCMatrix"),15)

}
