% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{pca_irlba_for_sparseMatrix}
\alias{pca_irlba_for_sparseMatrix}
\title{Run sparse PCA using irlba SVD}
\usage{
pca_irlba_for_sparseMatrix(x, n_comp, work = 3 * n_comp)
}
\arguments{
\item{x}{A sparse normalized matrix (features x cells)}

\item{n_comp}{The number of principal components to keep}
}
\value{
The rotated data, e.g. the cells x PC column in case of sc data.
}
\description{
Run sparse PCA using irlba SVD
}
