% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_differential_H1_scExp}
\alias{plot_differential_H1_scExp}
\title{Differential H1 distribution plot}
\usage{
plot_differential_H1_scExp(scExp_cf, cell_cluster = "C1")
}
\arguments{
\item{scExp_cf}{}

\item{cell_cluster}{}
}
\value{
A barplot of H1 distribution
}
\description{
Differential H1 distribution plot
}
\examples{
data("scExp")
scExp_cf = correlation_and_hierarchical_clust_scExp(scExp)
scExp_cf = choose_cluster_scExp(scExp_cf,nclust=2,consensus=FALSE)
scExp_cf = differential_analysis_scExp(scExp_cf)
plot_differential_H1_scExp(scExp_cf)
}
