% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_differential_summary_scExp}
\alias{plot_differential_summary_scExp}
\title{Differential summary barplot}
\usage{
plot_differential_summary_scExp(scExp_cf)
}
\arguments{
\item{scExp_cf}{A SingleCellExperiment object}
}
\value{
A barplot summary of differential analysis
}
\description{
Differential summary barplot
}
\examples{
data("scExp")
plot_differential_summary_scExp(scExp)
}
