% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_differential_volcano_scExp}
\alias{plot_differential_volcano_scExp}
\title{Volcano plot of differential features}
\usage{
plot_differential_volcano_scExp(
  scExp_cf,
  chromatin_group = "C1",
  cdiff.th = 1,
  qval.th = 0.01
)
}
\arguments{
\item{scExp_cf}{}

\item{chromatin_group}{}

\item{cdiff.th}{}

\item{qval.th}{}
}
\value{

}
\description{
Volcano plot of differential features
}
