% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_differential_volcano_scExp}
\alias{plot_differential_volcano_scExp}
\title{Volcano plot of differential features}
\usage{
plot_differential_volcano_scExp(
  scExp_cf,
  cell_cluster = "C1",
  cdiff.th = 1,
  qval.th = 0.01
)
}
\arguments{
\item{scExp_cf}{}

\item{cell_cluster}{}

\item{cdiff.th}{}

\item{qval.th}{}
}
\value{
A volcano plot of differential analysis of a specific cluster
}
\description{
Volcano plot of differential features
}
\examples{
data("scExp")
scExp_cf = correlation_and_hierarchical_clust_scExp(scExp)
scExp_cf = choose_cluster_scExp(scExp_cf,nclust=2,consensus=FALSE)
scExp_cf = differential_analysis_scExp(scExp_cf)
plot_differential_volcano_scExp(scExp_cf,"C1")
}
