% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_distribution_scExp}
\alias{plot_distribution_scExp}
\title{Plotting distribution of signal}
\usage{
plot_distribution_scExp(
  scExp,
  raw = TRUE,
  log10 = FALSE,
  pseudo_counts = 1,
  bins = 150
)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}

\item{raw}{Use raw counts ?}

\item{log10}{Transform using log10 ?}

\item{pseudo_counts}{Pseudo-count to add if using log10}

\item{bins}{Number of bins in the histogram}
}
\value{
A ggplot histogram representing the distribution of count per cell
}
\description{
Plotting distribution of signal
}
\examples{
data("scExp")
plot_distribution_scExp(scExp)
}
