% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_reduced_dim_scExp}
\alias{plot_reduced_dim_scExp}
\title{Plot reduced dimensions (PCA, TSNE, UMAP)}
\usage{
plot_reduced_dim_scExp(
  scExp,
  color_by = "sample_id",
  reduced_dim = c("PCA", "TSNE", "UMAP"),
  select_x = "Component_1",
  select_y = "Component_2",
  downsample = 5000,
  transparency = 0.6,
  size = 1,
  seed = 47
)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}

\item{color_by}{Feature used for coloration}

\item{reduced_dim}{Reduced Dimension used for plotting}

\item{select_x}{Which variable to select for x axis}

\item{select_y}{Which variable to select for y axis}

\item{downsample}{Number of cells to downsample}

\item{transparency}{Alpha parameter, between 0 and 1}

\item{seed}{Numeric random seed}
}
\value{
A ggplot geom_point plot of reduced dimension 2D reprensentation
}
\description{
Plot reduced dimensions (PCA, TSNE, UMAP)
}
\examples{
data("scExp")
plot_reduced_dim_scExp(scExp, color_by = "sample_id")
plot_reduced_dim_scExp(scExp, color_by = "total_counts")
plot_reduced_dim_scExp(scExp, reduced_dim = "UMAP")

}
