% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_reduced_dim_scExp}
\alias{plot_reduced_dim_scExp}
\title{Plot reduced dimensions (PCA, TSNE, UMAP)}
\usage{
plot_reduced_dim_scExp(
  scExp,
  color_by = "sample_id",
  reduced_dim = c("PCA", "TSNE", "UMAP"),
  select_x = "Component_1",
  select_y = "Component_2"
)
}
\arguments{
\item{scExp}{}

\item{color_by}{}

\item{reduced_dim}{}

\item{select_x}{}

\item{select_y}{}
}
\value{
A ggplot geom_point plot of reduced dimension 2D reprensentation
}
\description{
Plot reduced dimensions (PCA, TSNE, UMAP)
}
\examples{
data("scExp")
plot_reduced_dim_scExp(scExp, color_by = "sample_id")
plot_reduced_dim_scExp(scExp, color_by = "total_counts")
plot_reduced_dim_scExp(scExp, reduced_dim = "UMAP")
plot_reduced_dim_scExp(scExp, reduced_dim = "TSNE")

}
