% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{preprocess_CPM}
\alias{preprocess_CPM}
\title{Preprocess scExp - Counts Per Million (CPM)}
\usage{
preprocess_CPM(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}
}
\value{
A SingleCellExperiment object.
}
\description{
Preprocess scExp - Counts Per Million (CPM)
}
\examples{
raw <- create_scDataset_raw()
scExp = create_scExp(raw$mat, raw$annot)
scExp = preprocess_CPM(scExp)
head(SingleCellExperiment::normcounts(scExp))

}
