% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{preprocess_RPKM}
\alias{preprocess_RPKM}
\title{Preprocess scExp - Read per Kilobase Per Million (RPKM)}
\usage{
preprocess_RPKM(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}
}
\value{
A SingleCellExperiment object.
}
\description{
Preprocess scExp - Read per Kilobase Per Million (RPKM)
}
\examples{

scExp = create_scExp(create_scDataset_raw()$mat,  create_scDataset_raw()$annot)
scExp = preprocess_RPKM(scExp)
head(normcounts(scExp))
}
