% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{preprocess_TFIDF}
\alias{preprocess_TFIDF}
\title{Preprocess scExp - TF-IDF}
\usage{
preprocess_TFIDF(scExp, scale = 10000, log = TRUE)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}
}
\value{
A SingleCellExperiment object.
}
\description{
Preprocess scExp - TF-IDF
}
\examples{
scExp = create_scExp(create_scDataset_raw()$mat,create_scDataset_raw()$annot)
scExp = preprocess_TFIDF(scExp)
head(SingleCellExperiment::normcounts(scExp))

}
