% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{preprocess_TPM}
\alias{preprocess_TPM}
\title{Preprocess scExp - Transcripts per Million (TPM)}
\usage{
preprocess_TPM(scExp)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}
}
\value{
A SingleCellExperiment object.
}
\description{
Preprocess scExp - Transcripts per Million (TPM)
}
\examples{

raw <- create_scDataset_raw()
scExp = create_scExp(raw$mat, raw$annot)
scExp = preprocess_TPM(scExp)
head(SingleCellExperiment::normcounts(scExp))

}
