% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{reduce_dim_batch_correction}
\alias{reduce_dim_batch_correction}
\title{Reduce dimension with batch corrections}
\usage{
reduce_dim_batch_correction(scExp, mat, batch_list, n)
}
\arguments{
\item{scExp}{SingleCellExperiment}

\item{mat}{The normalized count matrix}

\item{batch_list}{List of batches}

\item{n}{Number of PCs to keep}
}
\value{
A list containing the SingleCellExperiment with batch info and
the corrected pca
}
\description{
Reduce dimension with batch corrections
}
