% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scExp}
\alias{scExp}
\title{A randomly generated SingleCellExperiment normalized and with reduced dimensions (hg38)}
\format{
scExp - a SingleCellExperiment with 300 cells and 600 features (genomic bins) in hg38:
\describe{
\item{chr}{A SingleCellExperiment}
}
}
\usage{
scExp
}
\description{
A randomly generated SingleCellExperiment normalized and with reduced dimensions (hg38)
}
\keyword{datasets}
