% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{separate_BAM_into_clusters}
\alias{separate_BAM_into_clusters}
\title{Separate BAM files into cell cluster BAM files}
\usage{
separate_BAM_into_clusters(affectation, odir, merged_bam)
}
\arguments{
\item{affectation}{An annotation data.frame containing cell_id and
cell_cluster columns}

\item{odir}{A valid output directory path}

\item{merged_bam}{A list of merged bam file paths}
}
\description{
Separate BAM files into cell cluster BAM files
}
