% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{separator_count_mat}
\alias{separator_count_mat}
\title{Determine Count matrix separator ("tab" or ",")}
\usage{
separator_count_mat(path_to_matrix)
}
\arguments{
\item{path_to_matrix}{A path towards the count matrix to check}
}
\value{
A character separator
}
\description{
Determine Count matrix separator ("tab" or ",")
}
