% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_filtering_reduction.R
\name{split_bam_file_into_single_cell_bams}
\alias{split_bam_file_into_single_cell_bams}
\title{split_bam_file_into_single_cell_bams}
\usage{
split_bam_file_into_single_cell_bams(
  bam_file,
  sample_id,
  outdir,
  min_coverage = 500,
  barcode_tag = "XB",
  verbose = TRUE
)
}
\arguments{
\item{bam_file}{BAM file}

\item{sample_id}{Name of the sample}

\item{outdir}{Output directory}

\item{min_coverage}{Minimum coverage to keep a cell (500)}

\item{barcode_tag}{Under which tag is the barcode information stored ? ("XB")}

\item{verbose}{Verbose (TRUE)}
}
\value{
Splitted single-cell Bam files
}
\description{
split_bam_file_into_single_cell_bams
}
