% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_analysis_gene_set_enrichment.R
\name{warning_DA}
\alias{warning_DA}
\title{Warning for differential_analysis_scExp}
\usage{
warning_DA(scExp, de_type, method, qval.th, logFC.th, block, group, ref)
}
\arguments{
\item{scExp}{A SingleCellExperiment object containing consclust with selected
number of cluster.}

\item{de_type}{Type of comparisons. Either 'one_vs_rest', to compare each
cluster against all others, or 'pairwise' to make 1 to 1 comparisons.
('one_vs_rest')}

\item{method}{Wilcoxon or edgerGLM}

\item{qval.th}{Adjusted p-value threshold. (0.01)}

\item{logFC.th}{Fold change threshold. (1)}

\item{block}{Use batches as blocking factors ?}

\item{group}{If de_type is custom, the group to compare (data.frame), must
be a one-column data.frame with cell_clusters or sample_id as character in
rows}

\item{ref}{If de_type is custom, the reference to compare (data.frame), must
be a one-column data.frame with cell_clusters or sample_id as character in
rows}
}
\value{
Warnings or Errors if the input are not correct
}
\description{
Warning for differential_analysis_scExp
}
