% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{warning_plot_reduced_dim_scExp}
\alias{warning_plot_reduced_dim_scExp}
\title{A warning helper for plot_reduced_dim_scExp}
\usage{
warning_plot_reduced_dim_scExp(
  scExp,
  color_by,
  reduced_dim,
  select_x,
  select_y
)
}
\arguments{
\item{scExp}{A SingleCellExperiment Object}

\item{color_by}{Feature used for coloration}

\item{reduced_dim}{Reduced Dimension used for plotting}

\item{select_x}{Which variable to select for x axis}

\item{select_y}{Which variable to select for y axis}
}
\value{
Warning or errors if the inputs are not correct
}
\description{
A warning helper for plot_reduced_dim_scExp
}
